+++ GorillaXS.Basics.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.8.34330.188
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "XSBasics", "XSBasics\XSBasics.csproj", "{1DAD65D5-D1D9-4646-8B56-4ADF03EA42B8}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "GorillaXS.Basics", "GorillaXS.Basics\GorillaXS.Basics.csproj", "{1DAD65D5-D1D9-4646-8B56-4ADF03EA42B8}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
