+++ GorillaXS.Basics/Directory.Build.props
