+++ GorillaXS.Basics/PluginInfo.cs
@@ -1,0 +1,9 @@
+﻿namespace GorillaXS.Basics
+{
+    internal class PluginInfo
+    {
+        public const string GUID = "luna.gorillaxs.basics";
+        public const string Name = "GorillaSX.Basics";
+        public const string Version = "1.0.0";
+    }
+}
