--- XSBasics/PluginInfo.cs
@@ -1,9 +1,0 @@
-﻿namespace XSBasics
-{
-    internal class PluginInfo
-    {
-        public const string GUID = "luna.xsbasics";
-        public const string Name = "XSBasics";
-        public const string Version = "1.0.0";
-    }
-}
